# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "WorkdirCreateAccepted": {
        "description": "POST /api/workdirs 的 202 受理响应：任务句柄 + 已登记的工作目录条目。",
        "properties": {
          "task_id": {
            "description": "导入任务句柄（GET /api/tasks/{id} 轮询）",
            "title": "Task Id",
            "type": "string"
          },
          "workdir": {
            "$ref": "#/components/schemas/WorkdirInfo",
            "description": "已登记（或复用）的工作目录条目"
          }
        },
        "required": [
          "task_id",
          "workdir"
        ],
        "title": "WorkdirCreateAccepted",
        "type": "object"
      },
      "WorkdirCreateRequest": {
        "description": "POST /api/workdirs 的请求体：登记工作目录（可选带初始导入）。\n\n带来源目录 = 复制导入（素材被复制进工作目录，初始导入）；不带 = 就地采用\n（直接使用目录内素材，不复制）。两种都会扫描 + 登记导入记录，均为长任务。",
        "properties": {
          "path": {
            "description": "工作目录绝对路径（服务端本地）",
            "title": "Path",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "原始素材目录；提供 = 复制导入，缺省 = 就地采用。来源须与工作目录不同且互不嵌套",
            "title": "Source"
          },
          "title": {
            "default": "",
            "description": "显示名；缺省 = 目录名",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "WorkdirCreateRequest",
        "type": "object"
      },
...
      "WorkdirInfo": {
        "description": "工作目录注册表条目——GET /api/workdirs 列表与详情的响应体。",
        "properties": {
          "id": {
            "description": "wid 短 ID（注册表主键，搬迁后不变）",
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "description": "最后使用时刻（Unix 秒，UTC）",
            "title": "Last Used At",
            "type": "number"
          },
          "path": {
            "description": "工作目录的规范绝对路径",
            "title": "Path",
            "type": "string"
          },
          "title": {
            "description": "显示名（默认 = 目录名，可改、允许重名）",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "path",
          "title",
          "last_used_at"
        ],
        "title": "WorkdirInfo",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "AI 打标工具（发图 + 指令产出 caption，支持迭代改写）",
    "title": "Dataset Factory",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workdirs": {
      "get": {
        "description": "列出全部登记的工作目录，按最后使用时间倒序（最近在前）。",
        "operationId": "list_all_api_workdirs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkdirInfo"
                  },
                  "title": "Response List All Api Workdirs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All",
        "tags": [
          "工作目录"
        ]
      },
      "post": {
        "description": "登记工作目录并受理初始导入（长任务，202 + task_id + Retry-After）。\n\n校验在登记之前完成（失败的请求不在注册表留痕）；注册表登记同步完成\n（wid 立即可用、幂等——同 realpath 只更新已有条目），扫描与复制走任务。",
        "operationId": "create_workdir_api_workdirs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkdirCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkdirCreateAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "工作目录或来源目录路径不合法（problem+json: workdir-path-invalid）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "来源目录与工作目录相同或互为嵌套（problem+json: import-source-conflict）"
          }
        },
        "summary": "Create Workdir",
        "tags": [
          "工作目录"
        ]
      }
    },
    "/api/workdirs/{wid}": {
      "delete": {
        "description": "确认路径后删除整个目录，失败返回残留位置并保留登记。",
        "operationId": "delete_registered_workdir_api_workdirs__wid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkdirDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Registered Workdir",
        "tags": [
          "工作目录"
        ]
      },
      "get": {
        "description": "按 wid 查注册表条目。",
        "operationId": "get_one_api_workdirs__wid__get",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkdirInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "wid 不在注册表（problem+json: workdir-not-found）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get One",
        "tags": [
          "工作目录"
        ]
      }
    },
...
    "/api/workdirs/{wid}/cleanup": {
      "post": {
        "description": "确认后清理选中的孤立产物，删除失败时返回暂存位置。",
        "operationId": "clean_selected_products_api_workdirs__wid__cleanup_post",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 9457 problem+json 错误体——二期新端点的统一错误形状。\n\n与一期 `ErrorDetail`（`{\"detail\"}`）并存：一期端点不返工，前端读 detail 兼容两者。\n扩展字段（occupier / 冲突清单等）由各错误在响应里按需附加，不进本模型。",
                  "properties": {
                    "detail": {
                      "description": "中文可操作消息——下一步该做什么",
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "description": "HTTP 状态码（与响应状态一致）",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "人读的短语概括",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "机器可读的错误类别 slug（如 task-not-found）",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 9457 problem+json 错误体——二期新端点的统一错误形状。\n\n与一期 `ErrorDetail`（`{\"detail\"}`）并存：一期端点不返工，前端读 detail 兼容两者。\n扩展字段（occupier / 冲突清单等）由各错误在响应里按需附加，不进本模型。",
                  "properties": {
                    "detail": {
                      "description": "中文可操作消息——下一步该做什么",
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "description": "HTTP 状态码（与响应状态一致）",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "人读的短语概括",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "机器可读的错误类别 slug（如 task-not-found）",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 9457 problem+json 错误体——二期新端点的统一错误形状。\n\n与一期 `ErrorDetail`（`{\"detail\"}`）并存：一期端点不返工，前端读 detail 兼容两者。\n扩展字段（occupier / 冲突清单等）由各错误在响应里按需附加，不进本模型。",
                  "properties": {
                    "detail": {
                      "description": "中文可操作消息——下一步该做什么",
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "description": "HTTP 状态码（与响应状态一致）",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "人读的短语概括",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "机器可读的错误类别 slug（如 task-not-found）",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clean Selected Products",
        "tags": [
          "工作目录"
        ]
      }
    },
...
    "/api/workdirs/{wid}/delete-preview": {
      "get": {
        "description": "删除前展示实际文件范围与原始素材警告。",
        "operationId": "get_deletion_preview_api_workdirs__wid__delete_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Deletion Preview",
        "tags": [
          "工作目录"
        ]
      }
    },
...
  }
}